v1.1: events cleanup
