+++ packages/core/base/src/types/events/internal/payloads.ts
@@ -1,0 +1,16 @@
+import { FiatEmbeddedCheckoutProps } from "@/types/embed";
+import { EmptyObject } from "@/types/system";
+
+import { CrossmintInternalEvents } from "./events";
+
+export interface CrossmintInternalEventMap {
+    [CrossmintInternalEvents.PARAMS_UPDATE]: ParamsUpdatePayload;
+    [CrossmintInternalEvents.UI_HEIGHT_CHANGED]: { height: number };
+    [CrossmintInternalEvents.CRYPTO_PAYMENT_USER_ACCEPTED]: { txId: string };
+    [CrossmintInternalEvents.CRYPTO_PAYMENT_USER_REJECTED]: EmptyObject;
+}
+
+// Params update
+export type ParamsUpdatePayload = Partial<
+    Record<keyof Omit<FiatEmbeddedCheckoutProps, "onEvent" | "environment">, any>
+>;
