+++ packages/ui/vanilla-ui/src/index.ts
@@ -1,4 +1,8 @@
 export * from "./components";
 
-export { CheckoutEvents, useCrossmintEvents } from "@crossmint/client-sdk-base";
-export type { CrossmintCheckoutEvent, CheckoutEventMap, CrossmintCheckoutEventUnion } from "@crossmint/client-sdk-base";
+export { CrossmintPublicEvents, useCrossmintEvents } from "@crossmint/client-sdk-base";
+export type {
+    CrossmintPublicEvent,
+    CrossmintPublicEventMap,
+    CrossmintPublicEventUnion,
+} from "@crossmint/client-sdk-base";
